OTHER REFACTOR Rename feature test files
